AI学术鉴伪/代码/AI_train/train_code/URN/requirements.txt
albumentations
baal
easydict
einops
numpy
opencv_python
Pillow
scikit_learn
scipy
timm
torchmetrics
torchsnooper
tqdm
